+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
+## [1.7.2] - 2021-08-06
+### Changed
+* Updated `com.unity.properties` to version `1.7.1-preview`.
+* Updated `com.unity.serialization` to version `1.7.3-preview`.
+
 ## [1.7.1] - 2021-06-07
 ### Changed
 * Updated `com.unity.serialization` to version `1.7.1-preview`.